clean: renamed types to core, removed core
