+++ core/game.go
@@ -1,8 +1,8 @@
 package core
 
 import (
-	"go_rts/types/geometry"
-	"go_rts/types/objects"
+	"go_rts/core/geometry"
+	"go_rts/core/objects"
 
 	"github.com/hajimehoshi/ebiten/v2"
 )
