+++ core/game_objects.go
@@ -1,9 +1,9 @@
 package core
 
 import (
-	"go_rts/types/geometry"
-	"go_rts/types/objects"
-	"go_rts/types/render"
+	"go_rts/core/geometry"
+	"go_rts/core/objects"
+	"go_rts/core/render"
 
 	"github.com/hajimehoshi/ebiten/v2"
 )
