+++ core/geometry/graph.go
