+++ core/geometry/point.go
