+++ core/geometry/point_test.go
