+++ core/geometry/position.go
