+++ core/geometry/position_test.go
