+++ core/geometry/rectangle.go
