+++ core/geometry/rectangle_test.go
