+++ core/input/interfaces.go
@@ -1,7 +1,7 @@
 package input
 
 import (
-	"go_rts/types/geometry"
+	"go_rts/core/geometry"
 
 	"github.com/ethereum/go-ethereum/event"
 	"github.com/hajimehoshi/ebiten/v2"
