+++ core/render/render.go
