+++ core/render/spritesheet.go
